Fix: Don't cache involved device ids
If a symlink is involved in a path and points to e.g. a device
that is not connected yet, caching the device id which would
return the id of the device that the symlink itself is located on/at,
would lead to the work item running even though the actual device
(that the symlink points to) is busy.